requirements.txt
aiogram~=2.12
environs~=8.0.0
Pillow~=8.3.1